django_scribe/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
decorator==5.1.1
dj-database-url==1.0.0
Django==4.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
executing==0.9.1
gunicorn==20.1.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tk==0.1.0
traitlets==5.3.0
urllib3==1.26.11
wcwidth==0.2.5
whitenoise==6.2.0
